+++ packages/ai/src/generate-object/stream-object.ts
@@ -510,7 +510,6 @@ class DefaultStreamObjectResult<
           maxRetries,
           headers,
           providerOptions,
-          abortSignal,
           output: outputStrategy.type as
             | 'object'
             | 'array'
@@ -561,7 +560,6 @@ class DefaultStreamObjectResult<
           modelId: model.modelId,
           providerOptions,
           headers,
-          abortSignal,
           functionId: telemetry?.functionId,
           promptMessages: callOptions.prompt,
         },